Change the order of repos used by modules
Details:
We use the quick start repos to test release candidates by having
these masquerade as artifacts that they serve to demonstrate.
Gradle looks for dependencies through repositories in the order in
which they are listed.
We put mavel local over all other repos to allow artifacts to be
published into mavenLocal and masqeuerade as existing versions
for testing.